Simplify project annotation dirs
