+++ src/projects/oraties/config/index.tsx
@@ -7,7 +7,7 @@ import {
 } from "../../../model/ProjectConfig";
 import { defaultConfig } from "../../default/config";
 import { defaultAnnotatedTextConfig } from "../../default/annotation/defaultAnnotatedTextConfig";
-import { OratiesMarker } from "../annotation/marker/OratiesMarker";
+import { OratiesMarker } from "../annotation/OratiesMarker";
 import { SearchItem } from "../SearchItem";
 import { dutchOratiesLabels } from "./dutchOratiesLabels";
 import { MetadataPanel } from "../MetadataPanel";
